# Changes

## Bot.js
- Renamed file to start.js
- Changed command and listener directories to [Commands](./Commands) and [Listeners](./Listeners/)

## embedCreator.js
- Renamed file to [EmbedCreator.js](./Functions/EmbedCreator.js)
- Added color presets
- Added send option
- Moved to [Functions](./Functions/)

## suggestionTable.js
- Renamed file to [suggestionDatabase.js](./Databases/suggestionDatabase.js)

## ./cmds/
- Renamed folder to [Commands](./Commands/)
- Moved every file to its own categories
- Made every file use [EmbedCreator](./Functions/EmbedCreator.js)

## ./Commands/Music
- Added music-related commands

## README.md
- Updated file

+++ Commands/Chill community-specific/Advertise.js
@@ -1,7 +1,6 @@
 const { Command } = require('discord-akairo');
 const color = require('img-color');
-const { adBoard, partnerships } = require('../logChannels.js');
-const createEmbed = require('../embedCreator.js');
+const createEmbed = require('../../Functions/EmbedCreator.js');
 
 class AdCommand extends Command {
   constructor() {
@@ -30,20 +29,19 @@ class AdCommand extends Command {
     });
   }
   async exec(message, args) {
+    if (message.guild.ownerID !== process.env.YWSSP) return;
     let discordInviteRegex = /https:\/\/discord\.gg\/[A-z0-9]{6,7}/.test(
       args.invite
     );
     //if the link is not a server invite or there isnt a link at all
     if (!discordInviteRegex) {
-      message.channel.send(
-        createEmbed({
-          message: message,
-          color: '#F44336',
-          title: 'Whoops!',
-          description: "That isn't a server invite!",
-          authorBool: true,
-        })
-      );
+      createEmbed(message, {
+        color: 'errorRed',
+        title: 'Whoops!',
+        description: "That isn't a server invite!",
+        authorBool: true,
+        send: 'channel',
+      });
       return message.delete();
     }
     const invite = await this.client.fetchInvite(args.invite);
@@ -53,9 +51,8 @@ class AdCommand extends Command {
     console.log(inviteGuildColor);
     if (!args.partner && !args.partner_claim) {
       if (!invite.temporary) {
-        this.client.channels.cache.get(adBoard).send(
-          createEmbed({
-            message: message,
+        this.client.channels.cache.get(process.env.ADVERTISEMENTS).send(
+          createEmbed(message, {
             color: `#${inviteGuildColor}`,
             title: invite.guild.name,
             thumbnail: invite.guild.iconURL(),
@@ -65,19 +62,17 @@ class AdCommand extends Command {
           })
         );
       } else {
-        message.channel.send(
-          createEmbed({
-            message: message,
-            color: '#F44336',
-            title: 'Whoops!',
-            description: 'Your invite is not temporary!',
-            authorBool: true,
-          })
-        );
+        createEmbed(message, {
+          color: 'errorRed',
+          title: 'Whoops!',
+          description: 'Your invite is not temporary!',
+          authorBool: true,
+          send: 'channel',
+        });
       }
     } else if (args.partner) {
       const memberIsVerified = message.member.roles.cache.some(
-        (role) => /Verified/.test(role.name) && !/\./.test(role.name)
+        (role) => /Verified/.test(role.name) && !/- .* -/.test(role.name)
       );
       if (memberIsVerified) {
         if (invite.memberCount > 30) {
@@ -85,8 +80,7 @@ class AdCommand extends Command {
             const ywssp = this.client.users.cache.get(process.env.YWSSP);
             await ywssp.send('New partnership!\nEmbed view:');
             await ywssp.send(
-              createEmbed({
-                message: message,
+              createEmbed(message, {
                 color: `#${inviteGuildColor}`,
                 title: invite.guild.name,
                 thumbnail: invite.guild.iconURL(),
@@ -99,57 +93,48 @@ class AdCommand extends Command {
             await ywssp.send(
               `y+advertise --partner-claim ${args.invite} ${args.description}`
             );
-            message.author.send(
-              createEmbed({
-                message: message,
-                title: 'Done!',
-                description:
-                  "The partnership isn't done yet, just wait for ywssp to see your request",
-                authorBool: true,
-              })
-            );
+            createEmbed(message, {
+              title: 'Done!',
+              description:
+                "The partnership isn't done yet, just wait for ywssp to see your request",
+              authorBool: true,
+              send: 'author',
+            });
           } else {
-            message.author.send(
-              createEmbed({
-                message: message,
-                color: '#F44336',
-                title: 'Whoops!',
-                description: 'Your invite is not permanent!',
-                authorBool: true,
-              })
-            );
+            createEmbed(message, {
+              color: 'errorRed',
+              title: 'Whoops!',
+              description: 'Your invite is not permanent!',
+              authorBool: true,
+              send: 'author',
+            });
           }
         } else {
-          message.author.send(
-            createEmbed({
-              message: message,
-              color: '#F44336',
-              title: 'Whoops!',
-              description: 'Your server has less than 30 members!',
-              authorBool: true,
-            })
-          );
+          createEmbed(message, {
+            color: 'errorRed',
+            title: 'Whoops!',
+            description: 'Your server has less than 30 members!',
+            authorBool: true,
+            send: 'author',
+          });
         }
       } else {
-        message.channel.send(
-          createEmbed({
-            message: message,
-            color: '#F44336',
-            title: 'Whoops!',
-            description: 'You need to be verified to request a partnership!',
-            authorBool: true,
-          })
-        );
+        createEmbed(message, {
+          color: 'errorRed',
+          title: 'Whoops!',
+          description: 'You need to be verified to request a partnership!',
+          authorBool: true,
+          send: 'channel',
+        });
       }
     } else if (args.partner_claim && message.author.id === process.env.YWSSP) {
       await this.client.channels.cache
-        .get(partnerships)
+        .get(process.env.partners)
         .send(
           `Hey <@&${message.guild.roles.everyone.id}>! Theres a new partnership!`
         );
-      await this.client.channels.cache.get(partnerships).send(
-        createEmbed({
-          message: message,
+      await this.client.channels.cache.get(process.env.partners).send(
+        createEmbed(message, {
           color: `#${inviteGuildColor}`,
           title: invite.guild.name,
           thumbnail: invite.guild.iconURL(),
